backend/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
asyncpg
alembic
passlib[argon2]
python-jose[cryptography]
python-dotenv
pydantic[email]
httpx
